Fix import paths to account for Login.tsx moving
