+++ client/src/stories/user-auth.pages.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react';
-import Login from '../pages/Login';
+import Login from '../pages/Login/Login';
 
 storiesOf('User Auth Pages | Login', module)
   .add('it renders', () => <Login/>)
